+++ tests/test_header.py
@@ -1,3 +1,4 @@
+from datetime import timedelta
 from distutils.version import StrictVersion
 
 import pytest
@@ -44,3 +45,36 @@ def test_strict_version():
     h.version_firmware = "v1.0.5"
 
     assert h.strict_version_firmware == StrictVersion("1.0.5")
+
+
+def test_timezone_conversion(dataset_master_simple):
+    header = dataset_master_simple[0].info
+    header.timezone = "Europe/Berlin"
+
+    berlin_start_time = header.local_datetime_start
+    berlin_stop_time = header.local_datetime_stop
+
+    assert berlin_start_time.hour - header.utc_datetime_start.hour == 1
+    assert berlin_stop_time.hour - header.utc_datetime_stop.hour == 1
+
+    header.timezone = "Europe/London"
+
+    london_start_time = header.local_datetime_start
+    london_stop_time = header.local_datetime_stop
+
+    assert london_start_time.hour - header.utc_datetime_start.hour == 0
+    assert london_stop_time.hour - header.utc_datetime_stop.hour == 0
+
+    assert berlin_start_time.hour - london_start_time.hour == 1
+    assert berlin_stop_time.hour - london_stop_time.hour == 1
+
+
+@pytest.mark.parametrize("attr", ("local_datetime_start", "local_datetime_stop"))
+def test_timezone_error(dataset_master_simple, attr):
+    header = dataset_master_simple[0].info
+
+    header.timezone = None
+    with pytest.raises(ValueError) as e:
+        getattr(header, attr)
+
+    assert "timezone" in str(e)