+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide section), and 
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [3.2] - 
+
+### Added
+- Nilspodlib is now timezone aware!
+  All loading methods for datasets, headers, and sessions now support a `tz` argument that is the string name of a valid
+  timezone (e.g. `Europe/Berlin`).
+  This allows access to the local start and end time in the header and a new local datetime index that can also be used 
+  when exporting data to pandas dataframes.
+  
+### Changed
+- The `utc_datetime_{start/end}` attributes of the header and the session are now properly reported in utc time instead 
+  of as naive datetime objects.
+- The `utc_datetime_counter` is now a pandas series. 
+
 # [3.1] - 11.01.2021
 
 ### Added